Rearrange org.flowable.cmmn.rest.service.api.history in subpackages Marked BaseSpringRestTestCase abstract to avoid being executed in the test suit
