+++ modules/flowable-cmmn-rest/src/main/java/org/flowable/cmmn/rest/service/api/CmmnRestResponseFactory.java
@@ -31,10 +31,10 @@ import org.flowable.cmmn.rest.service.api.engine.RestIdentityLink;
 import org.flowable.cmmn.rest.service.api.engine.variable.QueryVariable;
 import org.flowable.cmmn.rest.service.api.engine.variable.RestVariable;
 import org.flowable.cmmn.rest.service.api.engine.variable.RestVariable.RestVariableScope;
-import org.flowable.cmmn.rest.service.api.history.HistoricCaseInstanceResponse;
-import org.flowable.cmmn.rest.service.api.history.HistoricIdentityLinkResponse;
-import org.flowable.cmmn.rest.service.api.history.HistoricTaskInstanceResponse;
-import org.flowable.cmmn.rest.service.api.history.HistoricVariableInstanceResponse;
+import org.flowable.cmmn.rest.service.api.history.caze.HistoricCaseInstanceResponse;
+import org.flowable.cmmn.rest.service.api.history.task.HistoricIdentityLinkResponse;
+import org.flowable.cmmn.rest.service.api.history.task.HistoricTaskInstanceResponse;
+import org.flowable.cmmn.rest.service.api.history.variable.HistoricVariableInstanceResponse;
 import org.flowable.cmmn.rest.service.api.management.JobResponse;
 import org.flowable.cmmn.rest.service.api.repository.CaseDefinitionResponse;
 import org.flowable.cmmn.rest.service.api.repository.CmmnDeploymentResponse;
